DAY44/requirements.txt
qrcode
pillow
cryptography
